API BATCH ANALYSIS + FRONT BASE OF REWORK
TODO: AI TEST GEN BUGTEST, CI/CD GEN, CONN WITH REPO (AND PUSH), LIVE TESTING, CONN WITH JIRA, SLACK, GITHUB ACTIONS, GITLAB

+++ app/schemas.py
@@ -43,7 +43,7 @@ class ProjectOut(BaseModel):
     owner_id: int
     created_at: datetime
     updated_at: datetime
-    coverage: int = 0
+    coverage: float = 0.0
 
     model_config = ConfigDict(from_attributes=True)
